Improve plot scripts
As I am seeing very high values in the first measurements of each
batch, I have included std calculation to the output of the
processed csvs. I guess it has to do with using my laptop and
low cpu states, but let's see when using another computer.

Also, now the time for picto rendered views is averaged, so the
order in which we put the views rendering time is not relevant,
and the line grows in a uniform manner (way better for the
objective of the graph).