Include full process and plot script
The "process_and_plot.py" script takes the raw data from the
standalone and picto measurements and generates a single
plot that contains a subgraph per model of the measurements.

If desired, this script can also generate all itermediate
processed csvs (e.g. to check concrete values,
standardndard deviations, and so on)